Added levels class
