+++ levels/level2_1.txt
@@ -1,0 +1,24 @@
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 909 910 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 11 0 0 0 0 47 47 1 1 
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 35 0 0 0 0 47 47 1 1 
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 1 1 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 2 1 1 1 1 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 2 1 1 1 1 1 2 2 2 2 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 2 1 1 1 49 49 49 49 1 1 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 1 49 49 49 49 49 49 1 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 49 49 49 49 49 49 1 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 49 49 49 49 1 1 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 0 0 0 0 3 0 0 0 0 0 0 0 0 47 47 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 3 0 0 0 0 0 0 0 0 0 0 47 47 1 1 
+1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 1 
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 
+1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
